Allow alchemists to unlock transmutations
/api/transmutations/:name/unlock will unlock the given transmutation for
the current alchemist, if possible. Alchemists can unlock a
transmutation if it hasn't already been unlocked, they are the
appropriate level or have met the necessary transmutation requirements,
and have an unlock available for the transmutation's category.

Note that for level 4 Apprentice, tier two transmutations must be
completed before unlocking transmutations in higher tiers.